+++ Spotitube/src/main/java/domain/Track.java
@@ -1,13 +1,13 @@
 package domain;
 
 public abstract class Track {
-	private long id;
+	private int id;
 	private String performer;
 	private String title;
 	private String url;
 	private int duration;
 	
-	protected Track(long id, String performer, String title, String url, int duration)
+	protected Track(int id, String performer, String title, String url, int duration)
 	{
 		this.id = id;
 		this.performer = performer;
